micro_scalp_engine/ingest/kraken_fetcher/requirements.txt
google-cloud-pubsub
requests 